+++ app/src/main/java/camp/nextstep/edu/github/ui/SingleLiveEvent.kt
@@ -1,4 +1,4 @@
-package camp.nextstep.edu.github
+package camp.nextstep.edu.github.ui
 
 import androidx.annotation.MainThread
 import androidx.lifecycle.LifecycleOwner
